requirements.txt
Django==5.1.7
drf_spectacular_sidecar==2025.3.1
djangorestframework==3.15.2
psycopg2-binary==2.9.10
drf-spectacular==0.28.0
djangorestframework-simplejwt==5.5.0
bcrypt==4.3.0
psycopg2==2.9.10
python-dotenv
django-filter